# Make loop-idiom useful