@vchuravy 2021-12-09
Out of curiosity, what is the performance of `zeros` before and after?